+++ examples/synchronous_run.js
@@ -1,7 +1,7 @@
 /**
  * This example shows shows an act that has short runtime - just few seconds. It opens a webpage
  * http://goldengatebridge75.org/news/webcam.html that contains webcam stream from Golden Gate
- * bridge, takes a screenshot and saves it as output. This makes act executable at Apify platform
+ * bridge, takes a screenshot and saves it as output. This makes act executable on Apify platform
  * synchronously with a single request that also returns its output.
  *
  * Example is shared in library under https://www.apify.com/apify/example-golden-gate-webcam